Move common items to modulo_traits
