+++ modulo/src/core/mod.rs
@@ -1,2 +1,1 @@
-pub mod core_msg;
 pub mod core_thread;
