+++ modulo/src/file/mod.rs
@@ -1,3 +1,1 @@
-pub mod file_msg;
 pub mod file_thread;
-pub mod text;
