+++ modulo/src/lib.rs
@@ -1,5 +1,6 @@
 #[macro_use]
 extern crate log;
+extern crate modulo_traits;
 
 pub mod core;
 pub mod file;
