+++ modulo_traits/src/lib.rs
@@ -1,6 +1,3 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-    }
-}
+pub mod core_msg;
+pub mod file_msg;
+pub mod text;
